pyproject.toml
[project]
name = "stocks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "django-redis>=5.4.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.9.0",
    "yfinance>=0.2.40",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.3.0",
    "requests>=2.32.5",
]

production = [
    "sentry-sdk>=1.38.0",
    "django-cors-headers>=4.3.0",
]

# Ruff configuration
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E4",      # pycodestyle errors
    "E7",      # pycodestyle errors
    "E9",      # pycodestyle errors
    "F",       # Pyflakes
    "W",       # pycodestyle warnings
    "I",       # isort
    "N",       # pep8-naming
    "UP",      # pyupgrade
    "YTT",     # flake8-2020
    "S",       # flake8-bandit
    "BLE",     # flake8-blind-except
    "FBT",     # flake8-boolean-trap
    "B",       # flake8-bugbear
    "A",       # flake8-builtins
    "COM",     # flake8-commas
    "C4",      # flake8-comprehensions
    "DTZ",     # flake8-datetimez
    "T10",     # flake8-debugger
    "DJ",      # flake8-django
    "EM",      # flake8-errmsg
    "EXE",     # flake8-executable
    "FA",      # flake8-future-annotations
    "ISC",     # flake8-implicit-str-concat
    "ICN",     # flake8-import-conventions
    "G",       # flake8-logging-format
    "INP",     # flake8-no-pep420
    "PIE",     # flake8-pie
    "T20",     # flake8-print
    "PYI",     # flake8-pyi
    "PT",      # flake8-pytest-style
    "Q",       # flake8-quotes
    "RSE",     # flake8-raise
    "RET",     # flake8-return
    "SLF",     # flake8-self
    "SLOT",    # flake8-slots
    "SIM",     # flake8-simplify
    "TID",     # flake8-tidy-imports
    "TCH",     # flake8-type-checking
    "INT",     # flake8-gettext
    "ARG",     # flake8-unused-arguments
    "PTH",     # flake8-use-pathlib
    "TD",      # flake8-todos
    "FIX",     # flake8-fixme
    "ERA",     # eradicate
    "PD",      # pandas-vet
    "PGH",     # pygrep-hooks
    "PL",      # Pylint
    "TRY",     # tryceratops
    "FLY",     # flynt
    "NPY",     # NumPy-specific rules
    "AIR",     # Airflow
    "PERF",    # Perflint
    "FURB",    # refurb
    "LOG",     # flake8-logging
    "RUF",     # Ruff-specific rules
]

ignore = [
    # Django-specific ignores
    "DJ001",   # Avoid using `null=True` on string-based fields
    "DJ008",   # Model's `__str__` method should return a string

    # Common Django patterns that are acceptable
    "S101",    # Use of assert detected (common in Django tests)
    "S311",    # Standard pseudo-random generators are not suitable for security/cryptographic purposes

    # Style preferences
    "COM812",  # Trailing comma missing (conflicts with formatter)
    "COM819",  # Trailing comma prohibited (conflicts with formatter)
    "ISC001",  # Implicitly concatenated string literals on one line (conflicts with formatter)
    "Q000",    # Single quotes found but double quotes preferred (let formatter handle)
    "Q001",    # Single quote docstrings found but double quotes preferred
    "Q002",    # Docstring should use triple double quotes
    "Q003",    # Change outer quotes to avoid escaping inner quotes

    # Too strict for Django projects
    "FBT001",  # Boolean positional arg in function definition
    "FBT002",  # Boolean default value in function definition
    "ARG001",  # Unused function argument (common in Django views/signals)
    "ARG002",  # Unused method argument (common in Django methods)
    "TD002",   # Missing author in TODO
    "TD003",   # Missing issue link on the line following this TODO
    "FIX002",  # Line contains TODO, consider resolving the issue

    # Performance rules that might be too strict
    "PERF203", # `try`-`except` within a loop incurs performance overhead

    # Pylint rules that might be too strict for Django
    "PLR0913", # Too many arguments to function call
    "PLR2004", # Magic value used in comparison
    "PLW2901", # `for` loop variable overwritten by assignment target
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and print statements
"**/tests/**/*.py" = [
    "S101",    # asserts allowed in tests
    "ARG",     # Unused function args -> fixtures nevertheless are functionally relevant
    "FBT",     # Don't care about booleans as positional arguments in tests
    "PLR2004", # Magic value used in comparison
    "S311",    # Standard pseudo-random generators are not suitable for security/cryptographic purposes
    "T201",    # print statements allowed in tests
]

# Django settings files
"**/settings/**/*.py" = [
    "F405",    # may be undefined, or defined from star imports
    "F403",    # unable to detect undefined names
]

# Django migrations
"**/migrations/**/*.py" = [
    "E501",    # line too long (Django migrations can have long lines)
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
]

# Management commands
"**/management/commands/**/*.py" = [
    "T201",    # print statements allowed in management commands
    "ARG001",  # Unused function argument (Django command structure)
]

[tool.ruff.lint.isort]
# Django-specific isort configuration
known-first-party = ["users", "config"]
known-third-party = ["django", "rest_framework"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

# Django-specific configuration is handled through the DJ rules in select

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/text blocks and doctests are all supported.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
docstring-code-line-length = "dynamic"
